Require shallow clones to be requested explicitly
If you do a fetchTree on a Git repository, whether the result contains
a revCount attribute should not depend on whether that repository
happens to be a shallow clone or not. That would complicate caching a
lot and would be semantically messy. So applying fetchTree/fetchGit to
a shallow repository is now an error unless you pass the attribute
'shallow = true'. If 'shallow = true', we don't return revCount, even
if the repository is not actually shallow.

Note that Nix itself is not doing shallow clones at the moment. But it
could do so as an optimisation if the user specifies 'shallow = true'.

Issue #2988.

+++ src/libstore/fetchers/mercurial.cc
@@ -260,7 +260,7 @@ struct MercurialInput : Input
 
         Attrs infoAttrs({
             {"rev", input->rev->gitRev()},
-            {"revCount", revCount},
+            {"revCount", (int64_t) revCount},
         });
 
         if (!this->rev)
